requirements.txt
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
PyYAML==5.1.2
